Split lint and test jobs
This commit breaks out linting and testing into two separate, parallel
jobs. Since the RuboCop config specifies the target Ruby version (3.1),
there _shouldn't_ be a need to run the tool in each tested Ruby.

In this configuration, the linting job will run and report any errors in
parallel with the test job(s) defined in the matrix. The previous
addition that caches dependencies in the Ruby setup step should make all
of this run pretty quick.