Don't fail fast
This setting allows all jobs in the matrix to continue to completion
whether or not any of them fail along the way. This is most useful in
cases where, say, Ruby 3.3 fails on an actual error but we'd want to
know whether or not the same error appears in Ruby 3.2 and 3.1.

Absent this configuration, if the Ruby 3.3 error was encountered first,
all jobs would be cancelled without reporting success or failure of the
test suite.