Generate `.rubocop_todo.yml`
This commit streamlines the primary `.rubocop.yml` and, using `bundle
exec rubocop --auto-gen-config --auto-gen-only-exclude`, generates a
".rubocop_todo.yml" file that can serve as a guide for future
refactoring.

I've used this file with quite a bit of success on projects. A
reasonable goal is that, over time, the file is removed once all issues
have either been remediated or a sensible configuration change can be
applied to the main `.rubocop.yml` config.

Documentation:

https://docs.rubocop.org/rubocop/configuration.html#automatically-generated-configuration