Add issues and source code metadata to gemspec
The `gemspec` file supports a bunch of metadata [1]. Among that are a
couple of `*_uri` methods that RubyGems.org will helpfully show in the
sidebar of a page when viewing a gem (example [2]).

This commit adds two of those pieces of metadata:

- `bug_tracker_uri` pointing to the project's GitHub Issues page, and
- `source_code_uri` which dynamically links to the published version's
  source tree on GitHub.

The latter is most helpful when browsing versions of the gem on
RubyGems.org. Having a direct link to that version's tag on GitHub is
super handy.

[1] https://guides.rubygems.org/specification-reference/#metadata
[2] https://rubygems.org/gems/link-header-parser